+++ sdk/identity/identity/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Features Added
 
 - Added a new `requiredEnvVars` option to `DefaultAzureCredential` that can accept a single environment variable or an array of environment variables. All specified variables must be set (and non-empty) before credential instantiation. If any variable is missing or empty, an error is thrown listing all missing variables. [#35832](https://github.com/Azure/azure-sdk-for-js/pull/35832)
-- Introduced a new `DefaultEnvVars` enum-like type for environment variables that can be required by `DefaultAzureCredential` via the `requiredEnvVars` option.[#35832](https://github.com/Azure/azure-sdk-for-js/pull/35832)
+- Introduced a new `DefaultAzureCredentialEnvVars` enum-like type for environment variables that can be required by `DefaultAzureCredential` via the `requiredEnvVars` option.[#35832](https://github.com/Azure/azure-sdk-for-js/pull/35832)
 
 ### Breaking Changes